Remove a swathe of pointless tests and rename to reflect content
Coverage might drop as a result of this, but the tests don't test anything useful.
